+++ ext/sparse/sptab.c
@@ -148,7 +148,7 @@ static u_long sparse_table_hash(SparseTable *st, ScmObj key)
     if (!SCM_INTEGERP(r)) {
         Scm_Error("hash function %S returns non-integer: %S", h, r);
     }
-    return Scm_GetIntegerU(r);
+    return (u_long)Scm_GetIntegerU(r);
 }
 
 static int sparse_table_eq(SparseTable *st, ScmObj a, ScmObj b)
